+++ lib/messages.dart
@@ -2,6 +2,7 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:stripe/src/messages/converters.dart';
 
 part 'messages.g.dart';
+part 'src/messages/_message.dart';
 part 'src/messages/address.dart';
 part 'src/messages/balance_transaction.dart';
 part 'src/messages/charge.dart';