+++ lib/src/resources/payment_intent.dart
@@ -4,28 +4,27 @@ import 'package:logging/logging.dart';
 import 'package:stripe/messages.dart';
 
 import '../client.dart';
+import '_resource.dart';
 
 final log = Logger('Stripe PaymentIntentResource');
 
-class PaymentIntentResource {
-  final Client _client;
-  PaymentIntentResource(this._client);
+class PaymentIntentResource extends Resource<PaymentIntent> {
+  PaymentIntentResource(Client client) : super(client);
 
   Future<PaymentIntent> create(CreatePaymentIntentRequest request) async {
-    final response =
-        await _client.post('payment_intents', data: request.toJson());
+    final response = await post('payment_intents', data: request.toJson());
     return PaymentIntent.fromJson(response);
   }
 
   Future<PaymentIntent> retrieve(String paymentIntentId) async {
-    final map = await _client.get('payment_intents/$paymentIntentId');
+    final map = await get('payment_intents/$paymentIntentId');
     return PaymentIntent.fromJson(map);
   }
 
   /// Returns true if successful.
   Future<bool> cancel(String paymentIntentId) async {
     try {
-      await _client.post('payment_intents/$paymentIntentId/cancel');
+      await post('payment_intents/$paymentIntentId/cancel');
     } catch (e) {
       log.warning(e);
       return false;