+++ lib/src/resources/portal_session.dart
@@ -3,15 +3,15 @@ import 'dart:async';
 import 'package:stripe/messages.dart';
 
 import '../client.dart';
+import '_resource.dart';
 
-class PortalSessionResource {
-  final Client _client;
-  PortalSessionResource(this._client);
+class PortalSessionResource extends Resource<PortalSession> {
+  PortalSessionResource(Client client) : super(client);
 
   /// Creates a Stripe Checkout Session.
   Future<PortalSession> create(CreatePortalSessionRequest request) async {
     final response =
-        await _client.post('billing_portal/sessions', data: request.toJson());
+        await post('billing_portal/sessions', data: request.toJson());
     return PortalSession.fromJson(response);
   }
 }